# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Baseball Lineup Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /practices:
    get:
      operationId: listPractices
      tags: [practices]
      summary: List practice plans (most recent first), with attendance roll-ups
      responses:
        "200":
          description: Practices with attendance roll-ups
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PracticeSummary"
    post:
      operationId: createPractice
      tags: [practices]
      summary: Create a practice (blocks default to empty until coach generates or adds)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePracticeBody"
      responses:
        "201":
          description: Created practice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Practice"
  /practices/{id}:
    get:
      operationId: getPractice
      tags: [practices]
      summary: Get a practice with its blocks and per-player attendance
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Practice detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PracticeDetail"
        "404":
          description: Not found
    patch:
      operationId: updatePractice
      tags: [practices]
      summary: Update a practice's metadata or replace its blocks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePracticeBody"
      responses:
        "409":
          description: Row version conflict — the record changed on another device
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConflictError"
        "200":
          description: Updated practice
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Practice"
        "404":
          description: Not found
    delete:
      operationId: deletePractice
      tags: [practices]
      summary: Delete a practice (attendance cascades)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "409":
          description: Row version conflict — the record changed on another device
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConflictError"
        "204":
          description: Deleted
        "404":
          description: Not found
  /practices/{id}/attendance:
    put:
      operationId: replacePracticeAttendance
      tags: [practices]
      summary: Bulk-upsert per-player attendance for a practice
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReplaceAttendanceBody"
      responses:
        "200":
          description: Updated attendance
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PracticeAttendance"
        "404":
          description: Not found
...
components:
  schemas:
    ConflictError:
      type: object
      description: Returned with HTTP 409 when an If-Match row version no longer matches — another device changed the record. `current` is the row as it exists now.
      properties:
        error:
          type: string
        code:
          type: string
          enum: [row_version_conflict, idempotency_in_flight]
        current: {}
      required: [error, code]
...
    PracticeBlock:
      type: object
      description: One time-block within a practice plan (warmup, drill, scrimmage, etc.)
      properties:
        id:
          type: string
          description: Stable client-generated UUID for React keys + drag-reorder
        orderIndex:
          type: integer
          minimum: 0
        title:
          type: string
        durationMinutes:
          type: integer
          minimum: 1
          maximum: 240
        description:
          type: string
        drillType:
          type: string
          description: One of warmup | drill | scrimmage | conditioning | meeting
        focusAreas:
          type: array
          items:
            type: string
        groups:
          type: array
          description: Optional player groupings (defensive drills only — emitted by AI for infield/outfield/catching/pitching blocks).
          items:
            type: object
            properties:
              label:
                type: string
              playerNames:
                type: array
                items:
                  type: string
            required:
              - label
              - playerNames
        addressesFocusPoints:
          type: array
          description: Coach-authored focus points (verbatim strings from practice.focusPoints) this block targets. AI-tagged on generation.
          items:
            type: string
      required:
        - id
        - orderIndex
        - title
        - durationMinutes
        - description
        - drillType
        - focusAreas
    Practice:
      type: object
      properties:
        rowVersion:
          type: integer
          description: Optimistic-concurrency version, incremented on every update. Send back in If-Match on PATCH/DELETE to detect conflicts.
        id:
          type: integer
        date:
          type: string
          format: date-time
        durationMinutes:
          type: integer
        title:
          type: ["string", "null"]
        focusAreas:
          type: array
          items:
            type: string
        focusPoints:
          type: array
          description: Coach-authored "things to work on" — free-text symptoms/skills the AI designs drills around.
          items:
            type: string
        blocks:
          type: array
          items:
            $ref: "#/components/schemas/PracticeBlock"
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - date
        - durationMinutes
        - focusAreas
        - blocks
        - createdAt
        - updatedAt
    PracticeSummary:
      allOf:
        - $ref: "#/components/schemas/Practice"
        - type: object
          properties:
            attendanceMarked:
              type: integer
              description: Number of players with explicit attendance row (attended OR absent)
            attendancePresent:
              type: integer
              description: Number of players marked present
          required:
            - attendanceMarked
            - attendancePresent
    PracticeAttendance:
      type: object
      properties:
        rowVersion:
          type: integer
          description: Optimistic-concurrency version, incremented on every update. Send back in If-Match on PATCH/DELETE to detect conflicts.
        id:
          type: integer
        practiceId:
          type: integer
        playerId:
          type: integer
        attended:
          type: boolean
        notes:
          type: ["string", "null"]
        recordedAt:
          type: string
          format: date-time
      required:
        - id
        - practiceId
        - playerId
        - attended
        - recordedAt
    PracticeDetail:
      allOf:
        - $ref: "#/components/schemas/Practice"
        - type: object
          properties:
            attendance:
              type: array
              items:
                $ref: "#/components/schemas/PracticeAttendance"
          required:
            - attendance
    CreatePracticeBody:
      type: object
      properties:
        date:
          type: string
          format: date-time
        durationMinutes:
          type: integer
          minimum: 15
          maximum: 360
        title:
          type: ["string", "null"]
          maxLength: 120
        focusAreas:
          type: array
          items:
            type: string
        focusPoints:
          type: array
          items:
            type: string
            maxLength: 200
          maxItems: 20
        notes:
          type: ["string", "null"]
      required:
        - date
    UpdatePracticeBody:
      type: object
      properties:
        date:
          type: string
          format: date-time
        durationMinutes:
          type: integer
          minimum: 15
          maximum: 360
        title:
          type: ["string", "null"]
          maxLength: 120
        focusAreas:
          type: array
          items:
            type: string
        focusPoints:
          type: array
          description: Replace the practice's "things to work on" list. Pass [] to clear.
          items:
            type: string
            maxLength: 200
          maxItems: 20
        blocks:
          type: array
          items:
            $ref: "#/components/schemas/PracticeBlock"
        notes:
          type: ["string", "null"]
    AttendanceEntry:
      type: object
      properties:
        playerId:
          type: integer
        attended:
          type: boolean
        notes:
          type: ["string", "null"]
      required:
        - playerId
        - attended
    ReplaceAttendanceBody:
      type: object
      properties:
        entries:
          type: array
          items:
            $ref: "#/components/schemas/AttendanceEntry"
      required:
        - entries
...